Use Image.src to send data to loggly via Get request.
Since Loggly's API doesn't send cross-domain headers, the console gets
flooeded with an error message whenever we send data to the API via
actuall ajax POST or GET requests.

Loggly does, however, support a "tracking gif", which can be used to
send info.  With this, we'll set the src of an Image file to a url which
encapsulates the message we want to send.

The downside of this approach is that I no longer know how to test the
service.