Support toggling of logging levels.
This update adds a level() configuration method to the
LogglyLoggerProvider.  When setting a level on the logger, only messages
equal to or greater than that level will be sent to Loggly.

An example use-case is if you only want to send WARN messages to Loggly,
but continue to send INFO and DEBUG messages to the default $log
service.

This Fixes #1